functions/api/src/requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cognitojwt==1.4.1
cryptography==38.0.3
ecdsa==0.18.0
fastapi==0.86.0
h11==0.14.0
idna==3.4
jwt==1.3.1
mangum==0.16.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-jose==3.3.0
python-json-logger==2.0.7
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0